Update transaction Fee to be of type double

+++ src/shared/configuration/AppConfig.java
@@ -9,7 +9,7 @@ public class AppConfig
     private final BigDecimal startingBalance;
     private final int updateFrequencyInMs;
     private final int updateMaxVarianceInMs;
-    private final BigDecimal transactionFee;
+    private final double transactionFee;
     private final BigDecimal stockResetValue;
     private final int bankruptTimeInTicks;
     private final BigDecimal bankruptcyThreshold;
@@ -18,7 +18,7 @@ public class AppConfig
         this.startingBalance = BigDecimal.valueOf(1500);
         this.updateFrequencyInMs = 1000;
         this.updateMaxVarianceInMs = updateFrequencyInMs / 2;
-        this.transactionFee = BigDecimal.valueOf(25);
+        this.transactionFee = 25.0;
         this.stockResetValue = BigDecimal.valueOf(100);
         this.bankruptTimeInTicks = 5;
         this.bankruptcyThreshold = new BigDecimal("0.01");
@@ -41,7 +41,7 @@ public class AppConfig
         return updateFrequencyInMs;
     }
 
-    public BigDecimal getTransactionFee()
+    public double getTransactionFee()
     {
         return transactionFee;
     }
